Conflicts:
	.gitignore
	RootEngine/GUI/Source/guiInstance.cpp
	RootEngine/Network/Source/NetworkManager.cpp
	RootForce/Source/Main.cpp

+++ RootEngine/InputManager/Include/InputInterface.h
@@ -15,7 +15,6 @@ namespace RootEngine
 	namespace InputManager
 	{
 
-
 		class InputInterface abstract : public RootEngine::SubsystemInterface
 		{
 		public:
@@ -28,8 +27,8 @@ namespace RootEngine
 			virtual KeyState::KeyState GetKeyState(MouseButton::MouseButton p_button) = 0;
 
 			// Returns the global and delta mouse position
-			virtual glm::vec2 GetGlobalMousePos() = 0;
-			virtual glm::vec2 GetDeltaMousePos() = 0;
+			virtual glm::ivec2 GetGlobalMousePos() = 0;
+			virtual glm::ivec2 GetDeltaMousePos() = 0;
 
 		};
 	}
